pyproject.toml
[project]
name = "roux"
version = "0.1.0"
description = "Memory library for AI agents"
requires-python = ">=3.11"
dependencies = [
  "openai>=1.0.0",
  "python-dotenv>=1.0.0"
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"